+++ testproject/django_file_form_example/tests/live/test_without_ajax.py
@@ -1,5 +1,5 @@
-from django_file_form_example.base_live_testcase import BaseLiveTestCase
-from django_file_form_example.page import Page
+from django_file_form_example.tests.utils.base_live_testcase import BaseLiveTestCase
+from django_file_form_example.tests.utils.page import Page
 from django_file_form_example.models import Example, Example2