+++ testproject/django_file_form_example/views.py
@@ -5,11 +5,13 @@ from django.http import HttpResponseForbidden, HttpResponseRedirect
 from django.views import generic
 from django.urls import reverse
 from django.conf import settings
+from django.views.generic.base import ContextMixin
 from formtools.wizard.views import SessionWizardView
-from django_file_form.util import get_upload_path
+from django_file_form.util import get_upload_path, with_typehint
 from django_file_form.models import PlaceholderUploadedFile
 
 from . import forms
+from .models import Example
 
 
 def file_form_js():
@@ -20,9 +22,9 @@ def file_form_js():
 
 
 class BaseFormView(generic.FormView):
-    template_name = "example_form.html"
     use_ajax = True
     custom_js_files = ["example_form.js"]
+    template_name = "example_form.html"
 
     def get_success_url(self):
         return reverse("example_success")
@@ -32,9 +34,10 @@ class BaseFormView(generic.FormView):
         return super().form_valid(form)
 
     def get_context_data(self, **kwargs):
-        kwargs["use_ajax"] = self.use_ajax
         kwargs["custom_js_files"] = self.custom_js_files
         kwargs["file_form_js"] = file_form_js()
+        kwargs["use_ajax"] = self.use_ajax
+
         return super().get_context_data(**kwargs)
 
 
@@ -148,6 +151,30 @@ class WithCustomWidgetExample(PlaceholderView):
             ]
 
         return initial
+
+
+class FileModelFormMixin(with_typehint(ContextMixin)):
+    template_name = "example_form.html"
+    form_class = forms.ExampleModelForm
+    model = Example
+
+    def get_context_data(self, **kwargs):
+        kwargs["custom_js_files"] = ["example_form.js"]
+        kwargs["file_form_js"] = file_form_js()
+        kwargs["use_ajax"] = True
+
+        return super().get_context_data(**kwargs)
+
+    def get_success_url(self):
+        return reverse("example_success")
+
+
+class CreateModelFormView(FileModelFormMixin, generic.CreateView):
+    pass
+
+
+class EditModelFormView(FileModelFormMixin, generic.UpdateView):
+    pass
 
 
 def permission_denied(request, exception):