+++ testproject/testproject/settings_default.py
@@ -24,7 +24,9 @@ DEBUG = True
 
 DATABASES = dict(
     default=dict(
-        ENGINE="django.db.backends.dummy" if os.environ.get("CHECK_MIGRATIONS", "") == "true" else "django.db.backends.postgresql",
+        ENGINE="django.db.backends.dummy"
+        if os.environ.get("CHECK_MIGRATIONS", "") == "true"
+        else "django.db.backends.postgresql",
         NAME="django-file-form-example",
         USER="postgres",
         PASSWORD=os.environ.get("POSTGRES_PASSWORD", ""),